+++ bricks/dart_frog_prod_server/hooks/pre_gen.dart
@@ -2,14 +2,12 @@ import 'dart:async';
 import 'dart:io' as io;
 
 import 'package:dart_frog_gen/dart_frog_gen.dart';
-import 'package:mason/mason.dart' show HookContext;
+import 'package:mason/mason.dart'
+    show HookContext, defaultForeground, lightCyan;
 
 import 'src/create_bundle.dart';
 import 'src/exit_overrides.dart';
 import 'src/get_path_dependencies.dart';
-import 'src/report_external_path_dependencies.dart';
-import 'src/report_rogue_routes.dart';
-import 'src/report_route_conflicts.dart';
 
 typedef RouteConfigurationBuilder = RouteConfiguration Function(
   io.Directory directory,
@@ -37,9 +35,48 @@ Future<void> preGen(
     return exit(1);
   }
 
-  reportRouteConflicts(context, configuration, exit);
-  reportRogueRoutes(context, configuration, exit);
-  await reportExternalPathDependencies(context, projectDirectory, exit);
+  reportRouteConflicts(
+    configuration,
+    onRouteConflict: (
+      originalFilePath,
+      conflictingFilePath,
+      conflictingEndpoint,
+    ) {
+      context.logger.err(
+        '''Route conflict detected. ${lightCyan.wrap(originalFilePath)} and ${lightCyan.wrap(conflictingFilePath)} both resolve to ${lightCyan.wrap(conflictingEndpoint)}.''',
+      );
+    },
+    onViolationEnd: () {
+      exit(1);
+    },
+  );
+
+  reportRogueRoutes(
+    configuration,
+    onRogueRoute: (filePath, idealPath) {
+      context.logger.err(
+        '''Rogue route detected.${defaultForeground.wrap(' ')}Rename ${lightCyan.wrap(filePath)} to ${lightCyan.wrap(idealPath)}.''',
+      );
+    },
+    onViolationEnd: () {
+      exit(1);
+    },
+  );
+
+  await reportExternalPathDependencies(
+    projectDirectory,
+    onViolationStart: () {
+      context.logger
+        ..err('All path dependencies must be within the project.')
+        ..err('External path dependencies detected:');
+    },
+    onExternalPathDependency: (dependencyName, dependencyPath) {
+      context.logger.err('  \u{2022} $dependencyName from $dependencyPath');
+    },
+    onViolationEnd: () {
+      exit(1);
+    },
+  );
 
   context.vars = {
     'directories': configuration.directories
@@ -54,6 +91,7 @@ Future<void> preGen(
         : false,
     'serveStaticFiles': configuration.serveStaticFiles,
     'invokeCustomEntrypoint': configuration.invokeCustomEntrypoint,
+    'invokeCustomInit': configuration.invokeCustomInit,
     'pathDependencies': await getPathDependencies(projectDirectory),
   };
 }