+++ packages/dart_frog/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.3.6
+
+- fix: `Pipeline` does not maintain `RequestContext` ([#605](https://github.com/VeryGoodOpenSource/dart_frog/pull/605))
+- fix: `Response.json()` overwrites `content-type` header ([#596](https://github.com/VeryGoodOpenSource/dart_frog/pull/596))
+
+# 0.3.5
+
+- feat: add SecurityContext named argument to serve method
+- feat: add Response.stream
+
+# 0.3.4
+
+- feat: add support for `multipart/form-data` ([#551](https://github.com/VeryGoodOpenSource/dart_frog/pull/551))
+
 # 0.3.3
 
 - deps: upgrade to `Dart ">=2.19.0 <3.0.0"`