+++ packages/dart_frog_gen/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.4.0
+
+- feat!: re-land "add route configuration validation to gen" ([#614](https://github.com/VeryGoodOpenSource/dart_frog/pull/614))
+
+# 0.3.4
+
+- fix: revert "add route configuration validation to gen" ([#628](https://github.com/VeryGoodOpenSource/dart_frog/pull/628))
+
+# 0.3.3
+
+- feat: add route configuration validation to gen ([#614](https://github.com/VeryGoodOpenSource/dart_frog/pull/614))
+
+# 0.3.2
+
+- feat: detect custom init method ([#564](https://github.com/VeryGoodOpenSource/dart_frog/pull/564))
+
 # 0.3.1
 
 - deps: upgrade to `Dart ">=2.19.0 <3.0.0"`